Docker support & fix terraform template
Added a dockerfile that builds a docker image to use npk. This provides two benefits.

1. If CoalfireLabs decides to publish and maintain the npk image, then users can just run npk by using a single command line in docker.

2. If CoalfireLabs does not publish a docker image, users can still use this dockerfile to build the container themselves.

This prevents everyone from dealing with multiple terraform, python, pip and node versions that do not quite work together.

Pending to be done:

- Add ARGs, ENVs or a startup script (á la MySQL) so that the container can setup the aws credential profile when running and does not need to be a manual process.
- Add ARG, ENVs or a startup script (á la MySQL) so that the container can setup the /npk/terraform/npk-settings.json file and does not need to be a manual process.
- After these are done, we can remove `nano` (or any text editor) being installed, making the image smaller and faster to create.

This change also fixes the error in the `terraform/cognito_iam_roles.tf` template, which basically just had an extra comma. Fixes #49.